requirements.txt
# Discord bot runtime dependencies
discord.py>=2.4,<3.0
python-dotenv>=1.0,<2.0
playwright>=1.55,<2.0
tzdata>=2025.1,<2026.0
pytest>=8.3,<9.0
pytest-asyncio>=0.24,<1.0
pytest-mock>=3.14,<4.0
